Issue 63: Simplified AutoGsonAdapterGenerator by removing delegate
